Crate sdl2sketch

SDL2Sketch for Rust tries to simplify the use of rust-sdl2 by following the style of the p5.js API. It does not try to be a complete game engine, but just wants to make it as easy as possible to create visual applications in Rust without much boilerplate code. Code examples and the source code of SDL2Sketch can be found on its GitHub page.

Re-exports

pub use sdl2::pixels::Color;

pub use sdl2::keyboard::Keycode;

pub use sdl2::mouse::MouseButton;

pub use sdl2::render::Texture as Image;

Modules

utils
module containing utility functions

Structs

Sketch
This struct contains the necessary SDL2 subsystem objects and provides most of the API.

Enums

AngleMode
options for the interpretation of angle parameters

ImageMode
options for the interpretation of the parameters given to rect()

RectMode
options for the interpretation of the parameters given to rect()

Traits

MainLoopMethods
This trait must be implemented by the state struct of the application and provided to run().

Functions

run
starts the sketch and runs the main loop