Struct sdl2sketch::Sketch

pub struct Sketch { /* fields omitted */ }

This struct contains the necessary SDL2 subsystem objects and provides most of the API.

Methods

impl Sketch

pub fn new(width: u32, height: u32, title: &str) -> Self

create a new sketch

pub fn width(&self) -> i32

returns the width of the sketch in pixels

pub fn height(&self) -> i32

returns the height of the sketch in pixels

pub fn get_framerate(&mut self) -> f32

returns the current framerate in frames per second

In the p5.js API there is one function as getter and setter, framerate(), which has an optional argument.

pub fn set_framerate(&mut self, fps: u32)

sets the max. framerate in frames per second

max. setting 200 fps; In the p5.js API there is one function as getter and setter, framerate(), which has an optional argument.

pub fn no_loop(&mut self, setting: bool)

stops and restarts the main loop

In the p5.js API there are two functions for this, noLoop() and loop(). This does not work in rust since "loop" is a keyword, so a bool is needed as a parameter.

pub fn quit(&mut self)

exits the main loop

pub fn background(&mut self, color: Color)

clears the sketch by filling the whole sketch with the provided color

pub fn key_is_down(&self, code: Keycode) -> bool

checks if the key with the provided keycode is currently pressed

pub fn mouse_pos(&self) -> (i32, i32)

returns current x position of the mouse in pixel coordinates

In the p5.js API there are two state variables for this (mouseX and mouseY). If the mouse pointer is outside the sketch window, the function returns last position of the mouse inside the window(!)

pub fn mouse_is_pressed(&self) -> bool

checks if any mouse button is currently pressed

pub fn stroke(&mut self, color: Color)

After calling this function primitives will be drawn with an outline in the provided color.

pub fn no_stroke(&mut self)

After calling this function primitives will be drawn without outline.

pub fn fill(&mut self, color: Color)

After calling this function primitives will be drawn filled in the provided color.

pub fn no_fill(&mut self)

After calling this function primitives will be drawn without fill.

pub fn smooth(&mut self)

After calling this function primitives will be drawn with anti-aliasing. (nicer outline but slower)

pub fn no_smooth(&mut self)

After calling this function primitives will be drawn without anti-aliasing. (rugged outline but faster)

pub fn angle_mode(&mut self, mode: AngleMode)

After calling this function angle parameters of all subsequent function calls will be interpreted according to the provided mode.

pub fn rect_mode(&mut self, mode: RectMode)

After calling this function the parameters of all subsequent calls to rect() will be interpreted according to the provided mode.

pub fn image_mode(&mut self, mode: ImageMode)

After calling this function the parameters of all subsequent calls to image() will be interpreted according to the provided mode.

pub fn point(&mut self, x: i32, y: i32)

draws pixel-sized point at the provided coordinates

pub fn rect(&mut self, x: i32, y: i32, w: u32, h: u32)

draws a rectangle

pub fn line(&mut self, x1: i32, y1: i32, x2: i32, y2: i32)

draws a line

pub fn polygon(&mut self, vx: &[i32], vy: &[i32])

draws a polygon

SDL2-gfx API, not p5.js

pub fn triangle(&mut self, x1: i32, y1: i32, x2: i32, y2: i32, x3: i32, y3: i32)

draws a triangle

pub fn quad(
&mut self,
x1: i32,
y1: i32,
x2: i32,
y2: i32,
x3: i32,
y3: i32,
x4: i32,
y4: i32
)

draws a quad

pub fn arc(&mut self, x: i32, y: i32, r: u32, start: f32, end: f32)

draws arc (NOT COMPLETE)

zero for start/end angles is to the right (not on top)

TODO:

this is a circle arc not an ellipse arc (parameters are from SDL2-gfx API, not p5.js)

fill option not available

pub fn pie(&mut self, x: i32, y: i32, r: u32, start: f32, end: f32)

draws a pie

zero for start/end angles is to the right (not on top), there is no smooth option (ignores setting)

(SDL2-gfx API, not p5.js)

pub fn circle(&mut self, x: i32, y: i32, r: u32)

draws a circle

pub fn ellipse(&mut self, x: i32, y: i32, w: u32, h: u32)

draws an ellipse

pub fn image(&mut self, img: &Image, x: i32, y: i32, w: u32, h: u32)

displays an image at the given position (x,y) and size (w,h)

If w and/or h is 0 the original image width and/or height is used.