Module sdl2sketch::utils

module containing utility functions

Functions

constrain
utility function to constrain a value between low and high

deg_to_rad
utility function to convert angles from degrees to radians

hsv_to_rgb
utility function to convert a HSV color value to RGB (EXPERIMENTAL)

map
utility function to map a Float from one value range to another

norm
utility function to map a Float from one value range to [0,1]

rad_to_deg
utility function to convert angles from radians to degrees

rgb_to_hsv
utility function to convert a RGB color value to HSV (EXPERIMENTAL)