Trait sdl2sketch::MainLoopMethods

pub trait MainLoopMethods {
    fn setup(&mut self, _s: &mut Sketch) { ... }

    fn update(&mut self, _s: &mut Sketch) { ... }

    fn draw(&mut self, _s: &mut Sketch) { ... }

    fn key_pressed(&mut self, s: &mut Sketch, code: Keycode) { ... }

    fn key_released(&mut self, _s: &mut Sketch, _key: Keycode) { ... }

    fn mouse_moved(
        &mut self, 
        _s: &mut Sketch, 
        _x: i32, 
        _y: i32, 
        _xrel: i32, 
        _yrel: i32
    ) { ... }

    fn mouse_dragged(
        &mut self, 
        _s: &mut Sketch, 
        _x: i32, 
        _y: i32, 
        _xrel: i32, 
        _yrel: i32
    ) { ... }

    fn mouse_pressed(
        &mut self, 
        _s: &mut Sketch, 
        _button: MouseButton, 
        _x: i32, 
        _y: i32
    ) { ... }

    fn mouse_released(
        &mut self, 
        _s: &mut Sketch, 
        _button: MouseButton, 
        _x: i32, 
        _y: i32
    ) { ... }
}

This trait must be implemented by the state struct of the application and provided to run().

Provided Methods

fn setup(&mut self, _s: &mut Sketch)

called once before entering the main loop

fn update(&mut self, _s: &mut Sketch)

called every frame inside the main loop before draw()

fn draw(&mut self, _s: &mut Sketch)

called every frame inside the main loop

fn key_pressed(&mut self, s: &mut Sketch, code: Keycode)

called inside the main loop on a KeyDown event

The default implementation quits the main loop when Escape is pressed.

fn key_released(&mut self, _s: &mut Sketch, _key: Keycode)

called inside the main loop on a KeyUp event

fn mouse_moved(
&mut self,
_s: &mut Sketch,
_x: i32,
_y: i32,
_xrel: i32,
_yrel: i32
)

called inside the main loop on a MouseMotion event, if no MouseButton is pressed

fn mouse_dragged(
&mut self,
_s: &mut Sketch,
_x: i32,
_y: i32,
_xrel: i32,
_yrel: i32
)

called inside the main loop on a MouseMotion event, if a MouseButton is pressed

ignores which button(s) are pressed (TODO)

fn mouse_pressed(
&mut self,
_s: &mut Sketch,
_button: MouseButton,
_x: i32,
_y: i32
)

called inside the main loop on a MouseButtonDown event

fn mouse_released(
&mut self,
_s: &mut Sketch,
_button: MouseButton,
_x: i32,
_y: i32
)

called inside the main loop on a MouseButtonUp event